### 2025-08-04 #8742 @editorialbot
[PRE REVIEW]: emlearn-micropython: Machine Learning and Digital Signal Processing for MicroPython
**Submitting author:** @jonnor (<a href="http://orcid.org/0000-0002-0245-7456">Jon Nordby</a>)
**Repository:** https://github.com/emlearn/emlearn-micropython/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.8.0
