### 2025-08-13 3fa5419
CI: Try setup RISCV build for ESP32-C6
