### 2025-08-13 bd84916
CI: Disable C6 extmod build
Says it uses too much FLASH
