### 2025-08-13 a9b8309
emlearn_iir_q15: Remove custom memset
