### 2025-08-13 0b49d1e
Revert "emlearn_iir_q15: Remove custom memset"
This reverts commit a9b8309c77b0cc37580e3ec732fcfcdfcaf249e6.
