+++ src/emlearn_iir_q15/iir_filter.c
@@ -5,6 +5,16 @@
 
 #include <string.h>
 
+// memset is used by some standard C constructs
+#if !defined(__linux__)
+void *memcpy(void *dst, const void *src, size_t n) {
+    return mp_fun_table.memmove_(dst, src, n);
+}
+void *memset(void *s, int c, size_t n) {
+    return mp_fun_table.memset_(s, c, n);
+}
+#endif
+
 
 // MicroPython type for EmlIIR
 typedef struct _mp_obj_iir_filter_t {
