### 2025-08-13 34a9c6b
emlearn_iir_q15: Use weak linkage for memset and memcpy
