+++ src/emlearn_iir_q15/iir_filter.c
@@ -7,10 +7,10 @@
 
 // memset is used by some standard C constructs
 #if !defined(__linux__)
-void *memcpy(void *dst, const void *src, size_t n) {
+__attribute__((weak)) void *memcpy(void *dst, const void *src, size_t n) {
     return mp_fun_table.memmove_(dst, src, n);
 }
-void *memset(void *s, int c, size_t n) {
+__attribute__((weak)) void *memset(void *s, int c, size_t n) {
     return mp_fun_table.memset_(s, c, n);
 }
 #endif
