### 2025-08-13 f57bba4
emlearn_iir_q15: Try avoid duplicate symbol on RISC-V
