+++ src/emlearn_iir_q15/iir_filter.c
@@ -5,12 +5,13 @@
 
 #include <string.h>
 
-// memset is used by some standard C constructs
-#if !defined(__linux__)
-__attribute__((weak)) void *memcpy(void *dst, const void *src, size_t n) {
+// memset is used by some standard C constructs. Missing on some platforms, but not all...
+// Unfortunately cannot use weak symbols with mpy_ld
+#if !(defined(__linux__) || defined(__riscv__))
+void *memcpy(void *dst, const void *src, size_t n) {
     return mp_fun_table.memmove_(dst, src, n);
 }
-__attribute__((weak)) void *memset(void *s, int c, size_t n) {
+void *memset(void *s, int c, size_t n) {
     return mp_fun_table.memset_(s, c, n);
 }
 #endif
