### 2025-08-13 a5b67da
requirements: Add setuptools
Otherwise will get missing "distutils" on Python 3.12+
