+++ requirements.txt
@@ -2,3 +2,4 @@ emlearn>=0.21.2
 scikit-learn>=1.0.0
 ar>=1.0.0
 pyelftools>=0.31
+setuptools>=71.0.0
