### 2025-08-13 52ed3cf
CI: Add RISC-V as dependency for deploy step
