### 2025-08-13 #8742 @jonnor
Thank you @crvernon - this is my first submission to JOSS. I look forward to review feedback and will do my very best to address them :)
