### 2025-08-13 def779b
emlearn_iir_q15: Try other way to avoid memset on RISC-V
