### 2025-08-22 8d9db70
har_trees: Fix missing steps in README
