### 2025-08-22 80b27cd
har_trees: Add missing instructions
