### 2025-08-22 0c153a4
har_trees: Fix missing color_setup.py
