### 2025-09-20 81afab0
har_trees: Fix recorder on Zephyr, unhardcode number of columns
