### 2025-09-21 3870650
har_trees: Use pyproject.toml instead of requirements.txt
Allows to specify the subdirectory as a dependency
