+++ examples/har_trees/requirements.txt
@@ -1,10 +1,0 @@
-pandas
-pyarrow
-git+https://github.com/jonnor/leaf-clustering
-emlearn>=0.22.0
-scikit-learn
-setuptools
-structlog
-matplotlib
-mpremote
-pyyaml
