### 2025-10-04 2291498
linreg: Combine reference code
