+++ test_california_sklearn.py
@@ -1,262 +1,0 @@
-#!/usr/bin/env python3
-"""
-Reference implementation using scikit-learn ElasticNet for comparison with MicroPython module.
-Loads the same .npy files and computes MSE for benchmarking.
-"""
-
-import numpy as np
-from sklearn.linear_model import ElasticNet
-from sklearn.metrics import mean_squared_error, r2_score
-import time
-
-def load_data():
-    """Load the preprocessed California housing data."""
-    print("Loading data...")
-    X_train = np.load('X_train.npy')
-    X_test = np.load('X_test.npy')
-    y_train = np.load('y_train.npy')
-    y_test = np.load('y_test.npy')
-    
-    print(f"Train set: X={X_train.shape}, y={y_train.shape}")
-    print(f"Test set: X={X_test.shape}, y={y_test.shape}")
-    print(f"Data types: X={X_train.dtype}, y={y_train.dtype}")
-    
-    return X_train, X_test, y_train, y_test
-
-def test_elasticnet_configurations():
-    """Test different ElasticNet configurations to find good baselines."""
-    
-    X_train, X_test, y_train, y_test = load_data()
-    
-    # Test configurations: (alpha, l1_ratio, description)
-    configs = [
-        (0.0, 0.0, "No regularization (OLS)"),
-        (0.01, 0.0, "Ridge (alpha=0.01)"),
-        (0.01, 1.0, "LASSO (alpha=0.01)"),
-        (0.01, 0.5, "ElasticNet (alpha=0.01, l1_ratio=0.5)"),
-        (0.001, 0.5, "ElasticNet (alpha=0.001, l1_ratio=0.5)"),
-        (0.1, 0.5, "ElasticNet (alpha=0.1, l1_ratio=0.5)"),
-    ]
-    
-    print("\n" + "="*70)
-    print("ElasticNet Configuration Comparison")
-    print("="*70)
-    print(f"{'Configuration':<35} {'Train MSE':<12} {'Test MSE':<12} {'R²':<8} {'Time':<8}")
-    print("-"*70)
-    
-    results = []
-    
-    for alpha, l1_ratio, description in configs:
-        start_time = time.time()
-        
-        # Create and train model
-        if alpha == 0.0:
-            # Use regular linear regression for no regularization
-            from sklearn.linear_model import LinearRegression
-            model = LinearRegression()
-        else:
-            model = ElasticNet(alpha=alpha, l1_ratio=l1_ratio, max_iter=2000, random_state=42)
-        
-        model.fit(X_train, y_train)
-        
-        # Make predictions
-        y_train_pred = model.predict(X_train)
-        y_test_pred = model.predict(X_test)
-        
-        # Calculate metrics
-        train_mse = mean_squared_error(y_train, y_train_pred)
-        test_mse = mean_squared_error(y_test, y_test_pred)
-        test_r2 = r2_score(y_test, y_test_pred)
-        
-        elapsed_time = time.time() - start_time
-        
-        print(f"{description:<35} {train_mse:<12.6f} {test_mse:<12.6f} {test_r2:<8.3f} {elapsed_time:<8.3f}")
-        
-        results.append({
-            'config': description,
-            'alpha': alpha,
-            'l1_ratio': l1_ratio,
-            'train_mse': train_mse,
-            'test_mse': test_mse,
-            'r2': test_r2,
-            'time': elapsed_time,
-            'model': model
-        })
-    
-    return results
-
-def detailed_analysis_best_model(results):
-    """Perform detailed analysis on the best performing model."""
-    
-    # Find best model by test MSE
-    best_result = min(results, key=lambda x: x['test_mse'])
-    print(f"\n" + "="*50)
-    print("Detailed Analysis - Best Model")
-    print("="*50)
[…]
