+++ prepare_california.py
@@ -6,11 +6,17 @@ Saves scaled train/test splits as .npy files.
 """
 
 import numpy as np
+from sklearn.linear_model import ElasticNet
+from sklearn.metrics import mean_squared_error, r2_score
+import time
+
+
+import numpy as np
 from sklearn.datasets import fetch_california_housing
 from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import StandardScaler
 
-def prepare_california_housing_data():
+def prepare_california_housing_data(sample=None):
     """Download, preprocess and save California housing dataset."""
     
     print("Downloading California housing dataset...")
@@ -18,6 +24,11 @@ def prepare_california_housing_data():
     housing = fetch_california_housing()
     X, y = housing.data, housing.target
     
+    if sample is not None:
+        indices = np.random.choice(X.shape[0], size=sample, replace=False)
+        X = X[indices]
+        y = y[indices]
+
     print(f"Dataset shape: X={X.shape}, y={y.shape}")
     print(f"Features: {housing.feature_names}")
     print(f"Target: median house value in hundreds of thousands of dollars")
@@ -67,10 +78,246 @@ def prepare_california_housing_data():
 
 
 
-if __name__ == "__main__":
-    # Prepare the data
-    X_train, X_test, y_train, y_test = prepare_california_housing_data()
+def load_data():
+    """Load the preprocessed California housing data."""
+    print("Loading data...")
+    X_train = np.load('X_train.npy')
+    X_test = np.load('X_test.npy')
+    y_train = np.load('y_train.npy')
+    y_test = np.load('y_test.npy')
     
-    print("\nData preparation complete!")
-    print("Files ready for MicroPython testing:")
-    print("- X_train.npy, X_test.npy, y_train.npy, y_test.npy")
+    print(f"Train set: X={X_train.shape}, y={y_train.shape}")
+    print(f"Test set: X={X_test.shape}, y={y_test.shape}")
+    print(f"Data types: X={X_train.dtype}, y={y_train.dtype}")
+    
+    return X_train, X_test, y_train, y_test
+
+def test_elasticnet_configurations():
+    """Test different ElasticNet configurations to find good baselines."""
+    
+    X_train, X_test, y_train, y_test = load_data()
+    
+    # Test configurations: (alpha, l1_ratio, description)
+    configs = [
+        (0.0, 0.0, "No regularization (OLS)"),
+        (0.01, 0.0, "Ridge (alpha=0.01)"),
+        (0.01, 1.0, "LASSO (alpha=0.01)"),
+        (0.01, 0.5, "ElasticNet (alpha=0.01, l1_ratio=0.5)"),
+        (0.001, 0.5, "ElasticNet (alpha=0.001, l1_ratio=0.5)"),
+        (0.1, 0.5, "ElasticNet (alpha=0.1, l1_ratio=0.5)"),
+    ]
+    
+    print("\n" + "="*70)
+    print("ElasticNet Configuration Comparison")
+    print("="*70)
+    print(f"{'Configuration':<35} {'Train MSE':<12} {'Test MSE':<12} {'R²':<8} {'Time':<8}")
+    print("-"*70)
+    
+    results = []
+    
+    for alpha, l1_ratio, description in configs:
+        start_time = time.time()
+        
+        # Create and train model
+        if alpha == 0.0:
+            # Use regular linear regression for no regularization
+            from sklearn.linear_model import LinearRegression
+            model = LinearRegression()
+        else:
+            model = ElasticNet(alpha=alpha, l1_ratio=l1_ratio, max_iter=2000, random_state=42)
+        
+        model.fit(X_train, y_train)
+        
+        # Make predictions
[…]
