### 2025-10-05 fecb546
README: Mention linreg, and RISC-V works now
