### 2025-10-05 5d0779e
neighbors: Add cmake usermod build
