### 2025-10-05 1333a14
neighbors: Add support for building as usermod
