### 2025-10-05 ca6cd9b
neighbors: Avoid kwargs for predict()
Not needed and was buggy with usermod
