### 2025-10-16 emlearn/emlearn-micropython#49 @Puneetha-Ramachandra
[joss-review] Minor spelling error in paper
"The bottom plot slows " -> "The bottom plot shows "

@openjournals/joss-reviews#9093
