### 2025-10-16 emlearn/emlearn-micropython#50 @Puneetha-Ramachandra
[joss-review] Documentation Improvement for Installation and Testing Prerequisites
openjournals/joss-reviews#9093

Hi! I encountered a couple of minor hurdles while following the README for installation and testing. Clarifying these in the documentation would improve the setup experience for new users.

### 1. Undocumented Toolchain Dependency for Default Build
The default build command provided is for the `xtensawin` architecture:
```
make dist ARCH=xtensawin MPY_DIR=../micropython
```
**Problem**: This command fails with the error `make[1]: xtensa-esp32-elf-gcc: No such file or directory` if the required cross-compilation toolchain is not installed. While the prerequisites mention needing a "build toolchain," it's not immediately clear that the default example requires the specific ESP32 toolchain.

**Suggestion:** It would be helpful to either:

* Add a note specifying the required toolchain for the `xtensawin` example and link to the official setup instructions (e.g., from Espressif).

* Or, suggest building for a native architecture like `x64` as a simpler alternative for users who don't need to cross-compile.

### 2. Missing micropython Executable Prerequisite for Testing
**Problem**: After a successful build, running `make check` fails with the error `/bin/sh: line 1: micropython: command not found`. The current documentation instructs users to clone the MicroPython git repository, but this doesn't ensure the micropython executable is installed and available in the system's PATH. I was able to resolve this by running `sudo apt-get install micropython`.

**Suggestion**: Please update the prerequisites to explicitly state that the micropython command-line tool must be installed. Providing an example installation command for common package managers would be a great addition.
