### 2025-10-18 566de44
neighbors: Use m_del instead of m_free
