### 2025-10-26 emlearn/emlearn-micropython#51 @Puneetha-Ramachandra
[joss-review] Reproducing Figure 1 for har_trees
openjournals/joss-reviews#9093

My understanding is that the figure 1 in the paper comes from the har_trees example with the uci_har dataset. I was able to run the example with a few hiccups (mentioned below). I want to reproduce the probabilities plotted in the bottom graph of Figure 1. I could not find the instructions to do so.

**Can someone please guide me to reproduce these results?**

P.S: I ran into a few dep issues for running the example -
```
# mip doesn't exist in the micropython binary install
micropython -m mip install https://emlearn.github.io/emlearn-micropython/builds/master/x64_6.3/emlearn_trees.mpy
ImportError: no module named 'mip.__main__'
# for this I had to build the micropython unix port and use the micropython from the git repo

# Second, the micropython mip install doesn't install all the deps listed in emlearn-micropython/examples/har_trees/package.json. I had to manually do it
./build-standard/micropython -m mip install github:jonnor/micropython-mpu6886@master
./build-standard/micropython -m mip install github:jonnor/micropython-npyfile@master
./build-standard/micropython -m mip install github:jonnor/micropython-zipfile@master

```
It would be nice if the examples readme captured steps to install these deps.
