### 2025-10-27 68ee3d2
README: Move development instructions to separate file
So that it is less likely to be confused for usage instructions
