### 2025-10-27 58c7a6d
docs: Fix typo for module extension names
