+++ docs/getting_started_host.rst
@@ -57,7 +57,7 @@ Install emlearn-micropython modules
 ====================================
 
 emlearn-micropython is distributed as a set of MicroPython native modules.
-These are .npy file with native code, that can be installed at runtime using **mip**.
+These are .mpy file with native code, that can be installed at runtime using **mip**.
 This example uses the ``emlearn_trees`` module, so that is what we will install.
 
 .. code-block:: console
