### 2025-10-27 de17424
paper: Fix typo slows -> shows
Fixes #49
