### 2025-10-27 e0ef7ab
paper: Use Python script for creating plot
