### 2025-10-28 9496649
docs: Bit more details/links/notes about developing
