### 2025-10-28 79ac8b7
paper: Fix missing DOI for some references
