+++ paper/paper.bib
@@ -60,6 +60,7 @@
   booktitle={2012 16th international symposium on wearable computers},
   pages={108--109},
   year={2012},
+  doi={10.1109/iswc.2012.13},
   organization={IEEE}
 }
 
@@ -178,6 +179,7 @@ author={
   volume={27},
   pages={101778},
   year={2024},
+  doi={10.1016/j.softx.2024.101778},
   publisher={Elsevier}
 }
 
@@ -199,6 +201,7 @@ author={
   number={4},
   pages={1595--1623},
   year={2022},
+  doi={10.1016/j.jksuci.2021.11.019},
   publisher={Elsevier}
 }
 
