### 2025-10-30 972a816
paper: Update plotting code
Include the data exerpt in repo for ease of reproduction
