+++ paper/create_plot.py
@@ -5,6 +5,7 @@ import numpy
 import emlearn
 from sklearn.model_selection import cross_validate
 from sklearn.ensemble import RandomForestClassifier
+from sklearn.metrics import f1_score
 
 from plotting import plot_timeline, find_runs, configure_xaxis
 from plotting import make_label_colors, make_timeline_plot
@@ -27,6 +28,14 @@ colors = {
     'descending_stairs': 'rgb(252,205,229)',
 }
 
+sensor_columns = [
+    'hand_acceleration_6g_x',
+    'hand_acceleration_6g_y',
+    'hand_acceleration_6g_z',
+]
+
+def f1_macro(y_true, y_pred):
+    return f1_score(y_true, y_pred, average='macro')
 
 def train_and_run(data : pandas.DataFrame, class_columns, label_column='activity'):
     """
@@ -34,12 +43,6 @@ def train_and_run(data : pandas.DataFrame, class_columns, label_column='activity
 
     For a real example, see har_train.py in har_trees
     """
-
-    sensor_columns = [
-        'hand_acceleration_6g_x',
-        'hand_acceleration_6g_y',
-        'hand_acceleration_6g_z',
-    ]
 
     events = find_runs(data[label_column])
     sensor_data = convert_to_raw(data[sensor_columns])
@@ -54,7 +57,7 @@ def train_and_run(data : pandas.DataFrame, class_columns, label_column='activity
     X = combined[feature_columns]
     Y = combined[label_column]
 
-    # retrain on entire dataset (no test set or cross-validation)
+    # Train on entire dataset (no test set or cross-validation)
     # XXX: Not acceptable in real model, but here we are just looking to illustrate
     estimator.fit(X, Y)
 
@@ -64,9 +67,6 @@ def train_and_run(data : pandas.DataFrame, class_columns, label_column='activity
     converted.save(file=model_path, format='csv')
     print('Exported', model_path)
 
-    from sklearn.metrics import f1_score
-    def f1_macro(y_true, y_pred):
-        return f1_score(y_true, y_pred, average='macro')
     class_names = estimator.classes_
     ref_preds = pandas.DataFrame(estimator.predict_proba(X), index=X.index, columns=estimator.classes_)
     ref_score = f1_macro(Y, class_names[numpy.argmax(ref_preds, axis=1)])
@@ -75,10 +75,10 @@ def train_and_run(data : pandas.DataFrame, class_columns, label_column='activity
     out = process_data(sensor_data, model_path=model_path, classes=list(estimator.classes_))
     preds = out[estimator.classes_]
     c_score = f1_macro(Y, class_names[numpy.argmax(preds, axis=1)])
-    #print(preds.head())
 
     print('test scores', ref_score, c_score)
 
+    # Apply a bit of temporal smoothing on predictions
     rolling_window = 5
     smooth = preds[class_columns].rolling(rolling_window).median()
     smooth['activity'] =  make_label_track(smooth.index, events)
@@ -88,14 +88,10 @@ def train_and_run(data : pandas.DataFrame, class_columns, label_column='activity
 
 def main():
 
-    data_path = '../examples/har_trees/data/processed/pamap2.parquet'
+    data_path = 'pamap2_subject102_exerpt.parquet'
     plot_path = 'physical_activity_recognition_pipeline.png'
 
-    data = pandas.read_parquet(data_path)
-
-
-    # Select only a small subsection of data
-    one = data.loc['subject102']
+    sub = pandas.read_parquet(data_path)
 
     class_columns = [
         'Nordic_walking',
@@ -105,18 +101,15 @@ def main():
         #'running',
         'transient',
     ]
[…]
