+++ paper/README.md
@@ -1,0 +1,27 @@
+
+# emlearn-micropython JOSS paper
+
+## Example
+
+There is a plot in the paper.
+
+The MicroPython code is
+
+- [example.py](example.py)
+
+The following Python files are used for support:
+
+- [create_plot.py](create_plot.py)
+- [processing.py](processing.py)
+- [plotting.py](plotting.py)
+- [extract_data.py](extract_data.py)
+
+### Update plot
+
+Prerequisites: MicroPython for Unix installed.
+
+```
+python create_plot.py
+```
+
+Should update the file `physical_activity_recognition_pipeline.png`.
