### 2025-10-30 f6e6213
paper: Include instructions for dependencies
