+++ paper/README.md
@@ -3,9 +3,11 @@
 
 ## Example
 
-There is a plot in the paper.
+There is a plot in the paper,
+that demonstrates using emlearn_fft and emlearn_trees from emlearn-micropython
+to do Human Activity Recognition.
 
-The MicroPython code is
+The MicroPython code is found in:
 
 - [example.py](example.py)
 
@@ -16,9 +18,26 @@ The following Python files are used for support:
 - [plotting.py](plotting.py)
 - [extract_data.py](extract_data.py)
 
-### Update plot
+### Prerequities
 
-Prerequisites: MicroPython for Unix installed.
+- Python 3.10 or later
+- [MicroPython, Unix port](https://github.com/micropython/micropython/tree/master/ports/unix).
+
+### Install Python dependencies
+
+```
+pip install plotly numpy pandas scikit-learn
+```
+
+### Install MicroPython modules
+
+```
+micropython -m mip install github:jonnor/micropython-npyfile
+micropython -m mip install https://emlearn.github.io/emlearn-micropython/builds/master/x64_6.3/emlearn_trees.mpy
+micropython -m mip install https://emlearn.github.io/emlearn-micropython/builds/master/x64_6.3/emlearn_fft.mpy
+```
+
+### Run it
 
 ```
 python create_plot.py
