### 2025-10-30 ed7ab0c
CI: Also check paper example
