### 2025-10-30 5c9f665
paper: Add missing emlearn dependency
