+++ paper/README.md
@@ -26,7 +26,7 @@ The following Python files are used for support:
 ### Install Python dependencies
 
 ```
-pip install plotly numpy pandas scikit-learn
+pip install plotly numpy pandas scikit-learn emlearn
 ```
 
 ### Install MicroPython modules
