### 2025-10-30 1c67e14
docs: Update supported hardware for native modules
