+++ docs/native_modules.rst
@@ -20,12 +20,10 @@ The correct .mpy files to use depends on the CPU architecture of your microcontr
 Information is also available in the official documentation: `MicroPython: .mpy files <https://docs.micropython.org/en/latest/reference/mpyfiles.html#versioning-and-compatibility-of-mpy-files>`_.
 
 The following are **supported** as native modules at this time:
-``x64``, ``armv7m``, ``armv7emsp``, ``xtensawin``.
+``x64``, ``armv7m``, ``arm6vm`, ``armv7emsp``, ``xtensawin``, ``rv32imc``.
 
-The following are **not supported**: ``rv32imc``, ``armv6m``, ``xtensa``, ``x86``.
+The following are **not supported**: ``xtensa``, ``x86``.
 For these you will instead need to use :ref:`external_modules`.
-
-We hope to support ``arm6vm`` and ``rv32imc`` better in the future.
 There are no plans to support ``xtensa`` (ESP8266) or ``x86`` as native modules,
 as these are very old platforms.
 
