### 2025-10-30 b749e9a
paper: Update notes
