### 2025-10-30 4b93848
paper: Add missing setuptools dependency
