### 2025-10-30 4bc3a90
paper: Add missing pyarrow dependency
