### 2025-10-30 d1d2daa
paper: Fix missing install of MicroPython modules
