### 2025-10-30 8c51710
paper: Install Kaleido package
Needed for plotly PNG export
