### 2025-10-30 emlearn/emlearn-micropython#51 @jonnor
Since we only use a 15 minute section for the example, I decided to extract it and check it into git. This avoid downloading the entire dataset, which is over 1 GB. There is now steps-to-reproduce for the example in the paper at https://github.com/emlearn/emlearn-micropython/tree/master/paper#runnable-example
The Github actions run those steps successfully, so hopefully now it should be all good. Let me know how it goes :)
