### 2025-10-31 95db08c
docs: Add initial contributing guidelines
