+++ docs/CONTRIBUTING.md
@@ -1,0 +1,63 @@
+
+# Contributing
+
+Thanks for your interest in contributing!
+This is a volunteer-driven hobby project and we welcome contributions.
+
+## Ways to Contribute
+
+- Bug fixes
+- New features
+- Documentation improvements
+- Bug reports
+
+## Developing
+
+See [developing.md](./developing.md)
+
+## Guidelines for Pull Requests
+
+- Keep PRs focused on a single change
+- Write clear commit messages
+- Add tests for new functionality, or to cover against regressions
+- Remember to update documentation
+
+## Filing Issues
+
+Before creating an issue:
+- Search existing issues to avoid duplicates
+- Use a clear, descriptive title
+
+For bug reports, include:
+- Steps to reproduce
+- Expected vs actual behavior
+- Your environment (OS, version, etc.)
+
+For feature requests:
+- Describe the problem you're trying to solve / your usecase
+- Explain your proposed solution
+
+## AI/LLM Generated Code
+
+We accept the use of AI tools to assist with development,
+but all code must meet the same quality standards regardless of how it was created:
+
+- You must review and **understand** all code
+- Code must be tested and verified to work correctly
+- Follow the project's existing patterns and conventions
+- AI-generated code is not an excuse for lower quality
+- Code must be minimal to solve the problem at hand
+- Code must be simple, readable and easy to maintain
+
+You are responsible for any code you submit, regardless of its origin.
+Pull Requests that do not demonstrate sufficient attention to quality,
+will be rejected.
+
+## Code of Conduct
+
+Be respectful and constructive.
+We're all here to learn and build something useful together.
+
+## Questions?
+
+Ask in [discussions](https://github.com/emlearn/emlearn-micropython/discussions).
