### 2025-11-01 6026d38
docs: Cross-link developing and contribution
