### 2025-11-01 5e03198
README: Link directly to getting started
