### 2025-11-01 emlearn/emlearn-micropython#50 @jonnor
I have now improved the pointers to user documentation in README, and added separate contributing guidelines. So I hope it will be good now!
