### 2025-11-01 5cc8828
docs: Add API reference for emlearn_linreg
