+++ paper/notes.md
@@ -3,7 +3,6 @@
 
 Non-critical
 
-- Add linreg to the API reference
 - Add a documentation page about the wider MicroPython Data Science ecosystem?
 - Donate another 100 USD to JOSS
 
