### 2025-11-01 b4be8f8
docs: Mention how to build documentation
