### 2025-11-03 emlearn/emlearn-micropython#53 @mikey247
[JOSS-Review] Add instructions for running unit tests in /tests
https://github.com/openjournals/joss-reviews/issues/9093

Hi @jonnor,

For the checklist’s "Automated tests" item, could you document how to run the unit tests in /tests ? if it already exists, please point me to it.

Please include:

- Prerequisites: runtime/version, dependencies, environment variables

- Setup: install commands

- Run commands: all tests and a single test

- Expected output: what success/failure looks like

Suggest adding this to README or a TESTING.md and linking it from the main README.

Thanks.
