### 2025-11-03 emlearn/emlearn-micropython#50 @jonnor
Review was completed, closing this.
