### 2025-11-04 emlearn/emlearn-micropython#53 @mikey247
I ran into this issue when I ran the `make check` command. @jonnor
I've got micropython installed and followe the instructions in https://github.com/emlearn/emlearn-micropython/blob/master/docs/developing.md
