### 2025-11-04 emlearn/emlearn-micropython#53 @jonnor
Do you have a checkout of MicroPython at Desktop/micropython ? Per this comment in the docs:

> We assume that micropython git repository available. It is assumed to be at the same level as this repository in the file system. If using another location, adjust MPY_DIR accordingly. You should be using MicroPython 1.25 (or newer).

I will try to make that as well a bit more clear
