### 2025-11-04 emlearn/emlearn-micropython#53 @mikey247
hi @jonnor

I followed the instructions in https://github.com/micropython/micropython/blob/master/ports/unix/README.md to build Micropython and ran into this error running he `make` command. Note the warnings on the `make submodules` command; I'm not sure what to make of it.
