### 2025-11-10 emlearn/emlearn-micropython#53 @jonnor
Hmm. The first screenshot you shared had "pkg-config: file not found" errors, as output of the make submodules. Did you install it after that? Missing pkg-config and/or libffi would fit with the error that you get. But I do see an include path that has ffi in it in the failing build... But is from Mac OS / XCode though, not from homebrew - so maybe that is the problem.
You could try setting PKG_CONFIG_PATH to where homebrew has put the libffi .pc files? I unfortunately do not have a Mac for testing.

But, you seem to have installed `micropython` executable from homebrew at the lastest version. It should be possible to use that in combination with having a MicroPython source checkout (for the dynruntime.mk file that your first error got).
