### 2025-11-12 emlearn/emlearn-micropython#53 @mikey247
Yes, it was AFTER the first screenshot @jonnor

I think I've been able to fully setup Micropython, some test fail/skp but at last the libffi error is gone.

**Now back to emlearn, I now get this error. when I run `make check` command from [here](https://github.com/emlearn/emlearn-micropython/blob/master/docs/developing.md)

* "/Users/michaelumeokoli/Desktop/micropython/py/mpconfig.h:195:18: error: redefinition of typedef 'mp_int_t' is a C11 feature [-Werror,-Wtypedef-redefinition]"*
