### 2025-11-14 8e698c8
make: Try use a common .mk file to reduce copy-paste in modules
