+++ src/emlearn_trees/Makefile
@@ -1,19 +1,6 @@
-# Location of top-level MicroPython directory
-MPY_DIR = ../../micropython
-
-# Architecture to build for (x86, x64, armv6m, armv7m, xtensa, xtensawin)
-ARCH = x64
-
-# The ABI version for .mpy files
-MPY_ABI_VERSION := 6.3
 
 # Location of emlearn library
 EMLEARN_DIR := $(shell python3 -c "import emlearn; print(emlearn.includedir)")
-
-# enable linking of libm etc
-LINK_RUNTIME=1
-
-DIST_DIR := ../../dist/$(ARCH)_$(MPY_ABI_VERSION)
 
 # Name of module
 MOD = emlearn_trees
@@ -21,17 +8,8 @@ MOD = emlearn_trees
 # Source files (.c or .py)
 SRC = trees.c emlearn_trees.py
 
-# Releases
-DIST_FILE = $(DIST_DIR)/$(MOD).mpy
-$(DIST_DIR):
-	mkdir -p $@
-
-$(DIST_FILE): $(MOD).mpy $(DIST_DIR)
-	cp $< $@
-
 # Include to get the rules for compiling and linking the module
-include $(MPY_DIR)/py/dynruntime.mk
+include ../dynmodule.mk
 
+# Must come after dynmodule include
 CFLAGS += -I$(EMLEARN_DIR) -Wno-unused-function
-
-dist: $(DIST_FILE)
