### 2025-11-14 3f7c306
make: Allow passing CFLAGS_EXTRA through to module builds
