### 2025-11-14 1aec348
CI: Update to MicroPython 1.26.1
