### 2025-11-14 2438c5f
docs: Update to MicroPython 1.26
