### 2025-11-14 85bbead
CI: Try enable make check on Mac OS
