### 2025-11-14 23ac20d
Revert "CI: Try enable make check on Mac OS"
This reverts commit 85bbead84f60225e84f4abc635a2b11a38b7dd76.
