### 2025-11-14 61adcc8
CI: Mention issue preventing check to work on Mac OS
Ref https://github.com/micropython/micropython/issues/5500
