### 2025-11-14 82f2f91
docs: Mention how to run tests on host with external modules
This works also on Mac OS
