### 2025-11-14 c8a379f
tests: Add a test_all runner
