### 2025-11-14 53ca1cb
tests: Move executing behind __main__ check
So that we can import in test_all without tests starting at import time
