### 2025-11-14 dd087f4
tests: Use new test_all.py for check targets
