+++ Makefile
@@ -85,6 +85,9 @@ emlearn_arrayutils.results: $(MODULES_PATH)/emlearn_arrayutils.mpy
 emlearn_linreg.results: $(MODULES_PATH)/emlearn_linreg.mpy
 	MICROPYPATH=$(MODULES_PATH) $(MICROPYTHON_BIN) tests/test_linreg.py
 
+tests_all.results:
+	MICROPYPATH=$(MODULES_PATH) $(MICROPYTHON_BIN) tests/test_all.py
+
 $(PORT_DIR):
 	mkdir -p $@
 
@@ -95,11 +98,8 @@ $(UNIX_MICROPYTHON): $(PORT_DIR)
 unix: $(UNIX_MICROPYTHON)
 
 check_unix: $(UNIX_MICROPYTHON)
-	$(UNIX_MICROPYTHON) tests/test_trees.py
-	$(UNIX_MICROPYTHON) tests/test_iir.py
-	$(UNIX_MICROPYTHON) tests/test_fft.py
-	$(UNIX_MICROPYTHON) tests/test_arrayutils.py
-	echo SKIP $(UNIX_MICROPYTHON) tests/test_cnn.py
+	$(UNIX_MICROPYTHON) tests/test_all.py test_iir,test_fft,test_arrayutils
+	# TODO: enable more modules
 
 rp2: $(PORT_DIR)
 	make -C $(MPY_DIR)/ports/rp2 V=1 USER_C_MODULES=$(C_MODULES_SRC_PATH)/micropython.cmake FROZEN_MANIFEST=$(MANIFEST_PATH) CFLAGS_EXTRA='-Wno-unused-function -Wno-unused-function' -j4
@@ -130,7 +130,7 @@ release:
 	zip -r $(RELEASE_NAME).zip $(RELEASE_NAME)
 	#cp $(RELEASE_NAME).zip emlearn-micropython-latest.zip
 
-check: emlearn_trees.results emlearn_neighbors.results emlearn_iir.results emlearn_iir_q15.results emlearn_fft.results emlearn_kmeans.results emlearn_arrayutils.results emlearn_cnn.results emlearn_linreg.results
+check: tests_all.results
 
 dist: $(MODULES_PATH)/emlearn_trees.mpy $(MODULES_PATH)/emlearn_neighbors.mpy $(MODULES_PATH)/emlearn_iir.mpy $(MODULES_PATH)/emlearn_iir_q15.mpy $(MODULES_PATH)/emlearn_fft.mpy $(MODULES_PATH)/emlearn_kmeans.mpy $(MODULES_PATH)/emlearn_arrayutils.mpy $(MODULES_PATH)/emlearn_cnn_int8.mpy $(MODULES_PATH)/emlearn_cnn_fp32.mpy $(MODULES_PATH)/emlearn_linreg.mpy
 
