### 2025-11-14 4818875
tests: Improve test_all a bit
