### 2025-11-14 f5d9619
make: Use implicit rules instead of hardcoded lists for tests
