+++ Makefile
@@ -28,64 +28,39 @@ PORT_DIST_DIR=./dist/ports/$(PORT)/$(BOARD)
 
 UNIX_MICROPYTHON = ./dist/ports/unix/micropython
 
-$(MODULES_PATH)/emlearn_trees.mpy:
-	make -C src/emlearn_trees/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 clean dist
 
-$(MODULES_PATH)/emlearn_neighbors.mpy:
-	make -C src/emlearn_neighbors/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 clean dist
+# List of modules
+MODULES = emlearn_trees \
+	emlearn_neighbors \
+	emlearn_iir \
+	emlearn_fft \
+	emlearn_kmeans \
+	emlearn_iir_q15 \
+	emlearn_arrayutils \
+	emlearn_linreg \
+	emlearn_cnn_int8 \
+	emlearn_cnn_fp32
 
-$(MODULES_PATH)/emlearn_iir.mpy:
-	make -C src/emlearn_iir/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 clean dist
+# Generate list of .mpy files
+MODULE_MPYS = $(addprefix $(MODULES_PATH)/,$(addsuffix .mpy,$(MODULES)))
 
-$(MODULES_PATH)/emlearn_fft.mpy:
-	make -C src/emlearn_fft/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 clean dist
+# Special cases
+emlearn_cnn_int8_SRC = src/tinymaix_cnn
+emlearn_cnn_int8_CONFIG = CONFIG=int8
+emlearn_cnn_fp32_SRC = src/tinymaix_cnn
+emlearn_cnn_fp32_CONFIG = CONFIG=fp32
 
-$(MODULES_PATH)/emlearn_cnn_int8.mpy:
-	make -C src/tinymaix_cnn/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 CONFIG=int8 clean dist
+# Generate list of .mpy files
+MODULE_MPYS = $(addprefix $(MODULES_PATH)/,$(addsuffix .mpy,$(MODULES)))
 
-$(MODULES_PATH)/emlearn_cnn_fp32.mpy:
-	make -C src/tinymaix_cnn/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 CONFIG=fp32 clean dist
+# Build dynamic native module
+# defaults to
+$(MODULES_PATH)/%.mpy:
+	make -C $(or $($(*)_SRC),src/$*) \
+		ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} \
+		V=1 $($(*)_CONFIG) clean dist
 
-$(MODULES_PATH)/emlearn_kmeans.mpy:
-	make -C src/emlearn_kmeans/ ARCH=$(ARCH) MPY_DIR=$(MPY_DIR_ABS) CFLAGS_EXTRA=${CFLAGS_EXTRA} V=1 clean dist
-
-$(MODULES_PATH)/emlearn_iir_q15.mpy:
[…]
