### 2025-11-14 85bd94a
tests: Automatically add directories to sys.path
