### 2025-11-14 006ff02
docs: Mention how to run tests on device
