### 2025-11-14 323e452
docs: Mention expected output of tests
