### 2025-11-14 5d91d37
requirements: Also install mpremote
Used for running tests and other development tasks
