### 2025-11-16 8c48e73
har_trees: Start supporting custom features via user-specified program
