### 2025-11-16 0fcc50b
har_trees: Start moving window calculation to feature extractor
