+++ examples/har_trees/har_train.py
@@ -100,41 +100,6 @@ def evaluate(windows : pandas.DataFrame, groupby : str, hyperparameters : dict,
     return cv_results, estimator, splits, scores, figures
 
 
-def extract_windows(sensordata : pandas.DataFrame,
-    window_length : int,
-    window_hop : int,
-    groupby : list[str],
-    time_column = 'time',
-    ):
-
-    groups = sensordata.groupby(groupby, observed=True)
-
-
-    for group_idx, group_df in groups:
-
-        windows = []
-
-        # make sure order is correct
-        group_df = group_df.reset_index().set_index(time_column).sort_index()
-
-        # create windows
-        win_start = 0
-        length = len(group_df)
-        while win_start < length:
-            win_end = win_start + window_length
-            # ignore partial window at the end
-            if win_end > length:
-                break
-            
-            win = group_df.iloc[win_start:win_end].copy()
-            win['window'] = win.index[0]
-            assert len(win) == window_length, (len(win), window_length)
-   
-            windows.append(win)
-
-            win_start += window_hop
-
-        yield windows
 
 def assign_window_label(labels, majority=0.66):
     """
@@ -199,95 +164,106 @@ def timebased_features(windows : list[pandas.DataFrame],
     return df
 
 
-def custom_features(windows : list[pandas.DataFrame],
-        columns : list[str],
-        executable : str = '',
-        options : dict = {},
-        input_option : str = '--input',
-        output_option : str = '--output',
-        serialization : str = 'csv') -> pandas.DataFrame:
-    """
-    Run a program (executable) to compute features
+class DataProcessorProgram():
 
-    """
+    def __init__(self, executable : str,
+            options : dict = {},
+            input_option : str = '--input',
+            output_option : str = '--output',
+            serialization : str = 'csv'
+        ):
 
-    assert serialization == 'csv' # TODO: also support .npy
-    extension = serialization
+        self.executable = executable
+        self.options = options
+        self.input_option = input_option
+        self.output_option = output_option
+        self.serialization = serialization
 
-    # Filter columns
-    data = pandas.concat([ d for d in windows ])
+    def process(self, data : pandas.DataFrame) -> pandas.DataFrame:
+        """
+        Run a program (executable) to compute features
 
-    # FIXME: unhardcode
-    data['time'] = 0
-    data['gyro_x'] = 0
-    data['gyro_y'] = 0
-    data['gyro_z'] = 0
-    columns = ['time', 'acc_x', 'acc_y', 'acc_z', 'gyro_x', 'gyro_y', 'gyro_z']
+        Takes a DataFrame with sensor data as input.
+        The sensor data should be continous and regular in time.
+   
+        Returns windows with features.
+        The windows are usually overlapping in time.
+        """
 
-    data = data[columns]
+        assert self.serialization == 'csv' # TODO: also support .npy
+        extension = self.serialization
 
-    log.debug('custom-features-start', columns=list(data.columns))
+        # FIXME: unhardcode. Maybe allow specifying column_order ?
+        data['gyro_x'] = 0
+        data['gyro_y'] = 0
+        data['gyro_z'] = 0
+        data = data.reset_index()
+        ser_columns = ['time', 'acc_x', 'acc_y', 'acc_z', 'gyro_x', 'gyro_y', 'gyro_z']
+        data = data[ser_columns]
 
-    with tempfile.TemporaryDirectory() as tempdir:
-        data_path = os.path.join(tempdir, f'data.{extension}')
-        features_path = os.path.join(tempdir, f'features.{extension}')
+        log.debug('custom-features-start', columns=list(data.columns))
 
-        # Persist the data
-        data.to_csv(data_path, index=False)
+        with tempfile.TemporaryDirectory() as tempdir:
+            data_path = os.path.join(tempdir, f'data.{extension}')
[…]
