### 2025-11-16 0c137a3
har_trees: Bring back timebased features again
