+++ examples/har_trees/har_train.py
@@ -7,6 +7,7 @@ import tempfile
 import subprocess
 import json
 import itertools
+from typing import Optional
 
 import yaml
 import pandas
@@ -101,83 +102,29 @@ def evaluate(windows : pandas.DataFrame, groupby : str, hyperparameters : dict,
 
 
 
-def assign_window_label(labels, majority=0.66):
-    """
-    Assign the most common label to window, if it is above the @majority threshold
-    Otherwise return NA
-    """
-    counts = labels.value_counts(dropna=True)
-    threshold = majority * len(labels)
-    if counts.iloc[0] > threshold:
-        return counts.index[0]
-    else:
-        return None
-
-
-def timebased_features(windows : list[pandas.DataFrame],
-        columns : list[str],
-        python_bin='python') -> pandas.DataFrame:
-
-    #print('w', len(windows), columns)
-
-    here = os.path.dirname(__file__)
-    feature_extraction_script = os.path.join(here, 'compute_features.py')
-
-    data = numpy.stack([ d[columns] for d in windows ])
-    assert data.dtype == numpy.int16
-    assert data.shape[2] == 3, data.shape
-
-    #log.debug('data-range',
-    #    upper=numpy.quantile(data, 0.99),
-    #    lower=numpy.quantile(data, 0.01),
-    #)
-
-    with tempfile.TemporaryDirectory() as tempdir:
-        data_path = os.path.join(tempdir, 'data.npy')
-        features_path = os.path.join(tempdir, 'features.npy')
-
-        # Persist output
-        numpy.save(data_path, data)
-
-        # Run MicroPython program
-        args = [
-            python_bin,
-            feature_extraction_script,
-            data_path,
-            features_path,
-        ]
-        cmd = ' '.join(args)
-        #log.debug('run-timebased', cmd=cmd)
-        try:
-            out = subprocess.check_output(args)
-        except subprocess.CalledProcessError as e:
-            log.error('micropython-error', out=e.stdout, err=e.stderr)
-            raise e
-
-        # Load output
-        out = numpy.load(features_path)
-        assert len(out) == len(data)
-
-        # TODO: add feature names
-        df = pandas.DataFrame(out)
-
-    return df
-
-
 class DataProcessorProgram():
 
-    def __init__(self, executable : str,
+    def __init__(self, program : list[str],
             options : dict = {},
             input_option : str = '--input',
             output_option : str = '--output',
-            serialization : str = 'csv'
+            serialization : str = 'csv',
+            timeout : float = 10.0,
+            column_order : Optional[list[str]] = None,
         ):
 
-        self.executable = executable
+        self.program = program
         self.options = options
         self.input_option = input_option
         self.output_option = output_option
         self.serialization = serialization
+        self.timeout = timeout
+        self.column_order = column_order
+
+        supported = set(['npy', 'csv'])
+        if not serialization in supported:
+            raise ValueError(f'Unsupported serialization {serialization}. Valid: {supported}')
+
 
     def process(self, data : pandas.DataFrame) -> pandas.DataFrame:
         """
@@ -190,30 +137,33 @@ class DataProcessorProgram():
         The windows are usually overlapping in time.
         """
 
-        assert self.serialization == 'csv' # TODO: also support .npy
+
         extension = self.serialization
 
-        # FIXME: unhardcode. Maybe allow specifying column_order ?
-        data['gyro_x'] = 0
-        data['gyro_y'] = 0
-        data['gyro_z'] = 0
-        data = data.reset_index()
-        ser_columns = ['time', 'acc_x', 'acc_y', 'acc_z', 'gyro_x', 'gyro_y', 'gyro_z']
-        data = data[ser_columns]
-
-        log.debug('custom-features-start', columns=list(data.columns))
+        mod = data.copy()
[…]
