### 2025-11-17 8c1d188
har_trees: Try fix toothbrushing, support overlapped windows for timebased
