+++ examples/har_trees/har_train.py
@@ -178,7 +178,7 @@ class DataProcessorProgram():
 
             # Other options
             for k, v in self.options.items():
-                args += [ f'--{k}', v ]
+                args += [ f'--{k}', str(v) ]
 
             cmd = ' '.join(args)
             try:
@@ -194,11 +194,16 @@ class DataProcessorProgram():
                 # TODO: support feature names. Separat output file, with --features
                 out = numpy.load(features_path)
                 windows = pandas.DataFrame(out)
+                # FIXME: support reading times, not infer
                 span = (data.index.max() - data.index.min()).total_seconds()
-                dt = span / len(windows) # XXX: make correct
+                dt = span / len(windows)
+                log.debug('preprocess', windows=len(windows), dt=dt)
                 windows['time'] = dt * numpy.arange(len(windows))
             elif self.serialization == 'csv':
                 windows = pandas.read_csv(features_path)
+                span = (data.index.max() - data.index.min()).total_seconds()
+                dt = span / len(windows)
+                log.debug('preprocess', windows=len(windows), dt=dt)
             else:
                 raise NotImplementedError(self.serialization)
 
@@ -208,7 +213,7 @@ class DataProcessorProgram():
         time_in = data.index
         time_out = windows['time']
 
-        window_duration = pandas.Timedelta(5.0, unit='s') # XXX: hardcoded
+        window_duration = pandas.Timedelta(4.0, unit='s') # XXX: hardcoded
         start_delta = time_out.min() - time_in.min()
         assert abs(start_delta) <= window_duration, (start_delta, time_out.min(), time_in.min())
         end_delta = time_out.max() - time_in.max()
@@ -219,7 +224,7 @@ class DataProcessorProgram():
 class TimebasedFeatureExtractor(DataProcessorProgram):
 
     def __init__(self, python_bin='python', **kwargs):
-        super().__init__(self, input_option='', output_option='', serialization='npy', **kwargs)
+        super().__init__(self, serialization='npy', **kwargs)
 
         here = os.path.dirname(__file__)
         feature_extraction_script = os.path.join(here, 'compute_features.py')
@@ -234,6 +239,7 @@ def extract_features(sensordata : pandas.DataFrame,
     columns : list[str],
     groupby : list[str],
     extractor,
+    samplerate = 50,
     sensitivity = 2.0, # how many g range the int16 sensor data has
     label_column='activity',
     time_column='time',
@@ -300,12 +306,44 @@ def extract_features(sensordata : pandas.DataFrame,
     # for any time-dependent logic to stabilize, and to merge while ignoring the run-in
     def split_sections(data, groupby : list[str], time_column='time'):
         groups = sensordata.groupby(groupby, observed=True)
-        for group_idx, group_df in groups:
+        for group_idx, df in groups:
 
             # ensure sorted by time
-            group_df = group_df.reset_index().set_index(time_column).sort_index()
+            df = df.reset_index()
 
-            yield group_idx, group_df
+            # convert to time-delta, if neeeded
+            if pandas.api.types.is_datetime64_dtype(df[time_column]):
+                df[time_column] = df[time_column] - df[time_column].min() 
+
+            df = df.set_index(time_column).sort_index()
+
+            expected_freq = pandas.Timedelta(1/samplerate, unit='s')
+            diff = df.index.to_series().diff()
+            holes = diff[diff > expected_freq]
+            irregular = diff[diff != expected_freq].dropna()
+
+            # Convert to regular time-series
+            times = pandas.timedelta_range(df.index.min(), df.index.max(), freq=expected_freq)
[…]
