### 2025-11-20 emlearn/emlearn-micropython#53 @jonnor
Hi @mikey247
What is the difference between the two runs? Are you running that in a Python environment where you have the dependencies installed (where `pip install -r requirements.txt` was ran previously)?
