### 2025-11-22 emlearn/emlearn-micropython#53 @mikey247
pulled the updates and ran all the commands from scratch @jonnor
