### 2025-11-22 emlearn/emlearn-micropython#53 @jonnor
Hmm, in that case something must have failed in one of the earlier steps, that prevented the emlearn package from being installed. Can you please provide the complete log of all the steps?

We have a build-macos CI job which tests the same process for each commit. Ref https://github.com/emlearn/emlearn-micropython/blob/master/.github/workflows/build.yaml#L179
