### 2025-11-22 emlearn/emlearn-micropython#53 @mikey247
@jonnor I just followed the steps in https://github.com/emlearn/emlearn-micropython/blob/master/docs/developing.md, except for a new clone.
