### 2025-11-22 emlearn/emlearn-micropython#53 @jonnor
@mikey247 that is good. But something is failing, and I cannot debug or fix the failure if I do not see the output of the commands.
