+++ examples/har_trees/har_train.py
@@ -189,6 +189,11 @@ class DataProcessorProgram():
                     cmd=cmd, out=e.stdout, err=e.stderr, code=code)
                 raise e
 
+            log.debug('preprocessor-run',
+                cmd=cmd,
+                #out=out,
+            )
+
             # Load output
             if self.serialization == 'npy':
                 # TODO: support feature names. Separat output file, with --features
@@ -223,14 +228,24 @@ class DataProcessorProgram():
 
 class TimebasedFeatureExtractor(DataProcessorProgram):
 
-    def __init__(self, python_bin='python', **kwargs):
+    def __init__(self, sensitivity=4.0, python_bin='python', **kwargs):
         super().__init__(self, serialization='npy', **kwargs)
 
         here = os.path.dirname(__file__)
         feature_extraction_script = os.path.join(here, 'compute_features.py')
         self.program = [ python_bin, feature_extraction_script ]
 
+        self.sensitivity = sensitivity
+
     def process(self, data):
+
+        data = data.copy()
+        columns = self.column_order
+    
+        # Convert from floats in "g" to the sensor scaling in int16
+        data.loc[:, columns] = \
+            ((data.loc[:, columns] / self.sensitivity) * (2**15-1)).astype(numpy.int16)
+
         return super().process(data)
 
 
@@ -240,7 +255,6 @@ def extract_features(sensordata : pandas.DataFrame,
     groupby : list[str],
     extractor,
     samplerate = 50,
-    sensitivity = 2.0, # how many g range the int16 sensor data has
     label_column='activity',
     time_column='time',
     parallel_jobs : int = 4 ,
@@ -264,21 +278,8 @@ def extract_features(sensordata : pandas.DataFrame,
         # drop invalid data
         stream = stream.dropna(subset=columns)
 
-        # Convert from floats in "g" to the sensor scaling in int16
-        stream.loc[:, columns] = \
-            ((stream.loc[:, columns] / sensitivity) * (2**15-1)).astype(numpy.int16)
-
-        # FIXME: make sure time-series is regular
-        # potentially zero-fill ?
-
         # Extract features
         windows = extractor.process(stream)
-
-        # Convert features to 16-bit integers
-        # XXX: Assuming that they are already in resonable scale
-        # TODO: consider moving the quantization to inside timebased
-        feature_columns = list(set(windows.columns) - set([time_column]))
-        windows.loc[:,feature_columns] = windows[feature_columns].astype(numpy.int16)
 
         # Combine with identifying information
         # time should come from data processing
@@ -312,8 +313,8 @@ def extract_features(sensordata : pandas.DataFrame,
             df = df.reset_index()
 
             # convert to time-delta, if neeeded
-            if pandas.api.types.is_datetime64_dtype(df[time_column]):
-                df[time_column] = df[time_column] - df[time_column].min() 
+            #if pandas.api.types.is_datetime64_dtype(df[time_column]):
+            #    df[time_column] = df[time_column] - df[time_column].min() 
 
             df = df.set_index(time_column).sort_index()
 
@@ -401,8 +402,8 @@ def label_windows(sensordata,
         data = data.reset_index().set_index('time') # XXX: Why is this needed?
 
         # convert to time-delta, if neeeded
-        if pandas.api.types.is_datetime64_dtype(data.index):
-            data.index = data.index - data.index.min() 
+        #if pandas.api.types.is_datetime64_dtype(data.index):
+        #    data.index = data.index - data.index.min() 
 
[…]
