### 2025-11-30 f6847e9
ci: Try use XCode SDK 26.1 on Mac OS
