### 2025-11-30 ffa8f35
ci: Try -Wno-error
Since getting compile warnings as errors on Mac OS with SDK 26.1
Also skips unix standard build on Mac OS since it fails,
and is not needed as .mpy native modules are not implemented anyways.
