+++ .github/workflows/build.yaml
@@ -205,11 +205,11 @@ jobs:
     - name: Setup MicroPython X86
       working-directory: micropython
       run: |
-        source tools/ci.sh && ci_unix_32bit_setup && ci_unix_standard_build
-        mv ./ports/unix/build-standard/ ./ports/unix/build-nomodules/
+        make -C mpy-cross CFLAGS_EXTRA=-Wno-error
+        make -C ports/unix submodules
     - name: Build custom firmware with user modules, and tests. Unix/x64
       run: |
-        make check_unix V=1
+        make check_unix V=1 CFLAGS_EXTRA=-Wno-error
         mkdir -p ./dist/ports/macos
         mv ./dist/ports/unix/micropython ./dist/ports/macos/micropython
     - name: Build .mpy modules and run tests Unix/x64
