### 2025-11-30 605deee
ci: Try -Wno-error
Since getting compile warnings as errors on Mac OS with SDK 26.1
