### 2025-11-30 14468fc
CI: Try set -Wno-error also for standard build on Mac OS
