+++ .github/workflows/build.yaml
@@ -29,7 +29,7 @@ jobs:
     - name: Setup MicroPython X86
       working-directory: micropython
       run: |
-        source tools/ci.sh && ci_unix_32bit_setup && ci_unix_standard_build
+        CFLAGS_EXTRA='-Wno-error' source tools/ci.sh && ci_unix_32bit_setup && ci_unix_standard_build
         mv ./ports/unix/build-standard/ ./ports/unix/build-nomodules/
     - name: Build custom firmware with user modules, and tests. Unix/x64
       run: |
