### 2025-11-30 0dced66
CI: try skip unix standard build on Mac OS
