### 2025-11-30 87e16d1
CI: drop libffi on Mac OS
output says should already be installed
