+++ .github/workflows/build.yaml
@@ -191,7 +191,7 @@ jobs:
       run: sudo xcode-select -s /Applications/Xcode_26.1.1.app/Contents/Developer
 
     - name: Install OS dependencies
-      run:  brew install pkg-config libffi 
+      run:  brew install pkg-config 
     - uses: actions/setup-python@v5
       with:
         python-version: '3.10'
