### 2025-11-30 4619dcf
unix: Disable libffi explicitly
Was acting up for at least one person on Mac OS,
and we actually do not use it.
