### 2025-11-30 emlearn/emlearn-micropython#53 @jonnor
@mikey247 thanks for the full build log. I have no idea why libffi is not found. It should be a part of the XCode SDK... I have disabled it completely now. I have also updated our CI for Mac to use the Mac OS 26 SDK, which I saw is what you have. There I encountered some new build warnings that turned into errors, which I have documented a workaround for in the docs/developing.md - and with that I hope everything works...
Can you update to master and try again?
