### 2025-11-30 55a3850
plsr: Initial module code
