+++ tests/test_plsr.py
@@ -1,0 +1,244 @@
+
+import emlearn_plsr
+
+import array
+
+def assert_close(a, b, tolerance=0.1, name="value"):
+    """Simple assertion for floating point comparison"""
+    if abs(a - b) > tolerance:
+        raise AssertionError(f"{name}: expected {b}, got {a} (diff={abs(a-b)})")
+    print(f"  ✓ {name}: {a:.3f} ≈ {b:.3f}")
+
+def assert_true(condition, message):
+    """Simple assertion for boolean"""
+    if not condition:
+        raise AssertionError(message)
+    print(f"  ✓ {message}")
+
+def assert_equal(a, b, name="value"):
+    """Simple assertion for equality"""
+    if a != b:
+        raise AssertionError(f"{name}: expected {b}, got {a}")
+    print(f"  ✓ {name}: {a} == {b}")
+
+
+def test_simple_training():
+    """Test basic training and prediction"""
+    print("\n=== Test: Simple Training ===")
+    
+    # Simple data: y ≈ 2*x1 + 3*x2 + 1*x3
+    X = array.array('f', [
+        1, 0, 0,      # y = 2
+        0, 1, 0,      # y = 3
+        0, 0, 1,      # y = 1
+        1, 1, 0,      # y = 5
+        1, 0, 1,      # y = 3
+        0, 1, 1,      # y = 4
+        1, 1, 1,      # y = 6
+        0.5, 0.5, 0.5 # y = 3
+    ])
+    y = array.array('f', [2, 3, 1, 5, 3, 4, 6, 3])
+    
+    # Create and train
+    model = emlearn_plsr.new(8, 3, 2)
+    success = emlearn_plsr.fit(model, X, y, max_iterations=100, tolerance=1e-6, verbose=0)
+    
+    assert_true(success, "Training succeeded")
+    assert_true(model.is_complete(), "All components trained")
+    
+    # Test predictions
+    x_test = array.array('f', [1, 0, 0])
+    y_pred = model.predict(x_test)
+    assert_close(y_pred, 2.0, 0.2, "Prediction x=[1,0,0]")
+    
+    x_test = array.array('f', [0, 1, 0])
+    y_pred = model.predict(x_test)
+    assert_close(y_pred, 3.0, 0.2, "Prediction x=[0,1,0]")
+    
+    x_test = array.array('f', [2, 1, 0.5])
+    y_pred = model.predict(x_test)
+    assert_close(y_pred, 7.5, 0.5, "Prediction x=[2,1,0.5]")
+    
+    print("  ✓ All predictions within tolerance")
+
+
+def test_stepwise_training():
+    """Test step-by-step training API"""
+    print("\n=== Test: Stepwise Training ===")
+    
+    # Same simple data
+    X = array.array('f', [
+        1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0,
+        1, 0, 1, 0, 1, 1, 1, 1, 1, 0.5, 0.5, 0.5
+    ])
+    y = array.array('f', [2, 3, 1, 5, 3, 4, 6, 3])
+    
+    # Create model
+    model = emlearn_plsr.new(8, 3, 2)
+    
+    # Start training
+    model.fit_start(X, y)
+    assert_true(not model.is_complete(), "Not complete initially")
+    
+    # Train components
+    components_trained = 0
+    while not model.is_complete():
+        # Iterate until convergence
+        iterations = 0
+        while not model.is_converged() and iterations < 100:
+            model.step(1e-6)
+            iterations += 1
+        
+        assert_true(model.is_converged(), f"Component {components_trained} converged")
+        assert_true(iterations < 100, f"Component {components_trained} converged in time")
+        
+        # Finalize
+        model.finalize_component()
[…]
