+++ tests/test_all.py
@@ -30,6 +30,7 @@ TEST_MODULES=[
     'test_linreg_california',
     'test_logreg',
     'test_logreg_cancer',
+    'test_plsr',
     'test_neighbors',
     'test_trees',
     'test_extratrees',
