+++ src/emlearn_plsr/eml_plsr.h
@@ -1,0 +1,508 @@
+/**
+ * @file eml_plsr.h
+ * @brief Embedded Machine Learning - Partial Least Squares Regression (NIPALS)
+ * 
+ * Header-only implementation using single precision floats and a single
+ * pre-allocated memory block.
+ * 
+ * Based on: Wold, H. (1966) "Estimation of principal components and related 
+ * models by iterative least squares"
+ * 
+ * Usage:
+ *   // Default: implementation is included (header-only mode)
+ *   #include "eml_plsr.h"
+ *   
+ *   // To disable implementation in some files:
+ *   #define EML_PLSR_IMPLEMENTATION 0
+ *   #include "eml_plsr.h"
+ */
+
+#ifndef EML_PLSR_H
+#define EML_PLSR_H
+
+#include <stdint.h>
+#include <stdbool.h>
+#include <stddef.h>
+
+/**
+ * @brief Calculate required memory size for PLS regression
+ * 
+ * @param n_samples Number of training samples
+ * @param n_features Number of input features
+ * @param n_components Number of PLS components
+ * @return Required memory size in bytes
+ */
+#define EML_PLSR_MEMORY_SIZE(n_samples, n_features, n_components) \
+    (((size_t)(n_samples) * (size_t)(n_features) + /* inputs_work */ \
+      (size_t)(n_samples) +                         /* targets_work */ \
+      (size_t)(n_features) * (size_t)(n_components) + /* weights */ \
+      (size_t)(n_features) * (size_t)(n_components) + /* loadings_x */ \
+      (size_t)(n_components) +                      /* loadings_y */ \
+      (size_t)(n_samples) * (size_t)(n_components) + /* scores */ \
+      (size_t)(n_samples) +                         /* score_curr */ \
+      (size_t)(n_samples) +                         /* score_y_curr */ \
+      (size_t)(n_features) +                        /* weight_curr */ \
+      (size_t)(n_features)                          /* loading_x_curr */ \
+     ) * sizeof(float))
+
+/* Error codes */
+typedef enum _EmlError {
+    EmlOk = 0,
+    EmlSizeMismatch,
+    EmlUnsupported,
+    EmlUninitialized,
+    EmlPostconditionFailed,
+    EmlUnknownError,
+    EmlErrors,
+} EmlError;
+
+/**
+ * @brief PLS regression state structure
+ */
+typedef struct {
+    /* Problem dimensions */
+    uint16_t n_samples;
+    uint16_t n_features;
+    uint16_t n_components;
+    
+    /* Pointers into user-provided memory */
+    float *inputs_work;
+    float *targets_work;
+    float *weights;
+    float *loadings_x;
+    float *loadings_y;
+    float *scores;
+    float *score_curr;
+    float *score_y_curr;
+    float *weight_curr;
+    float *loading_x_curr;
+    
+    /* Training state */
+    uint16_t current_component;
+    uint16_t current_iter;
+    bool component_converged;
+    float convergence_metric;
+} eml_plsr_t;
+
+/* ============================================================================
+ * IMPLEMENTATION CONTROL
+ * ============================================================================
+ */
+
+#ifndef EML_PLSR_IMPLEMENTATION
+#define EML_PLSR_IMPLEMENTATION 1
+#endif
+
+#if !EML_PLSR_IMPLEMENTATION
+/* Function declarations (only when implementation is disabled) */
+
+EmlError eml_plsr_init(
+    eml_plsr_t *plsr,
[…]
