### 2025-11-30 4926f6c
plsr: Clean up unused code
