### 2025-11-30 d2642fc
plsr: Try use __builtin_sqrtf
