### 2025-11-30 8d99552
plsr: Use different sqrtf depending on ARCH
