### 2025-11-30 f1a4af6
plsr: Use different sqrtf depending on ARCH
