### 2025-11-30 cd28fd1
plsr: Some examples on real-world data
