+++ spectrofood_download.py
@@ -1,0 +1,133 @@
+import os
+import pandas as pd
+import numpy as np
+import urllib.request
+from io import StringIO
+
+import pandas as pd
+import numpy as np
+from sklearn.model_selection import train_test_split
+from sklearn.preprocessing import StandardScaler
+from sklearn.cross_decomposition import PLSRegression
+from sklearn.metrics import mean_squared_error, r2_score
+
+
+DATA_URL = "https://zenodo.org/records/8362947/files/SpectroFood_dataset.csv?download=1"
+
+def download_dataset(data_dir):
+    os.makedirs(data_dir, exist_ok=True)
+    csv_file = os.path.join(data_dir, "SpectroFood_dataset.csv")
+    if not os.path.exists(csv_file):
+        print("Downloading SpectroFood CSV...")
+        urllib.request.urlretrieve(DATA_URL, csv_file)
+    return csv_file
+
+
+def load_spectrofood_chunks(csv_file, target_col="dry_matter", food_col="food"):
+    """
+    Splits CSV into chunks using empty lines (newlines) as separators.
+    Each chunk is loaded with pandas.read_csv separately.
+    Returns list of tuples: (food_name, DataFrame)
+    """
+    chunks = []
+    with open(csv_file, 'r') as f:
+        content = f.read()
+
+    # Split into raw text blocks on empty lines
+    raw_chunks = [c.strip() for c in content.split("\n\n") if c.strip()]
+    # FIXME: only returns 1 chunk right now
+    print(len(raw_chunks))
+
+    for chunk_text in raw_chunks:
+        # Use StringIO to read the chunk as CSV
+        chunk_io = StringIO(chunk_text)
+        try:
+            df_chunk = pd.read_csv(chunk_io, dtype=str, keep_default_na=False)
+        except pd.errors.EmptyDataError:
+            continue  # skip empty chunks
+
+        # Determine food name: use the first column of the first row
+        if food_col in df_chunk.columns:
+            food_name = df_chunk[food_col].iloc[0].strip().replace(" ", "_")
+        else:
+            food_name = str(df_chunk.iloc[0, 0]).strip().replace(" ", "_")
+
+        # Convert numeric columns to float, ignore errors
+        df_chunk = df_chunk.apply(pd.to_numeric, errors='coerce')
+        chunks.append((food_name, df_chunk))
+
+    return chunks
+
+def preprocess_chunk(df_chunk, target_col="DRY MATTER"):
+    """
+    Converts DataFrame to C-contiguous X and y numpy arrays
+    """
+
+    #print(df_chunk.columns)
+
+    # Keep only rows where the target column is numeric
+    df_chunk = df_chunk[pd.to_numeric(df_chunk[target_col], errors='coerce').notna()].copy()
+
+    # Drop columns that are entirely NaN
+    df_chunk = df_chunk.dropna(axis=1, how='all')
+
+    # Drop rows that are entirely NaN
+    df_chunk = df_chunk.dropna(axis=0, how='any')
+
+    exclude_cols = [c for c in df_chunk.columns if c == target_col or df_chunk[c].dtype == object]
+    X = df_chunk.drop(columns=exclude_cols).values.astype(np.float32)
+    y = df_chunk[target_col].values.astype(np.float32).reshape(-1, 1)
+
+    # Standardize
+    scaler_X = StandardScaler()
+    #scaler_y = StandardScaler()
+    X = scaler_X.fit_transform(X)
+    #y = scaler_y.fit_transform(y)
+
+    X = np.ascontiguousarray(X)
+    y = np.ascontiguousarray(y)
+    return X, y
+
[…]
