### 2025-12-01 emlearn/emlearn-micropython#53 @mikey247
Hey @jonnor, Ive been able to get the tests to work.

Issue: MicroPython build failed with linker error Undefined symbols for architecture arm64: "_mp_module_ffi"

probable cause: maybe Stale/corrupted build artifacts in the build-standard directory. The build system had inconsistent state - possibly from an interrupted previous build, incomplete compilation, or corrupted intermediate files.

Solution: Clean the build directory to remove all artifacts, then rebuild:

```
../micropython/ports/unix
make clean
cd -
make check_unix

```
