### 2025-12-04 emlearn/emlearn-micropython#60 @swouf
[joss-review] Error while trying to reproduce results from paper
https://github.com/openjournals/joss-reviews/issues/9093#issuecomment-3355248465
Hi, while trying to reproduce the results from the paper, I encountered an error and I don't find my way around it. I applied the steps for the [runnable example](https://github.com/emlearn/emlearn-micropython/tree/master/paper#runnable-example), but I can't manage to get around this. I have the following error.

Could someone help me with this? That would be great.

```bash
vscode ➜ /workspaces/python/emlearn-micropython/paper (master) $ python create_plot.py
missing values 0.0033332962967078145 %
window duration 256 2.56 s
run-cmd micropython /workspaces/python/emlearn-micropython/paper/example.py /tmp/tmppsdbb0u8/input.npy /tmp/tmppsdbb0u8/output.npy
Traceback (most recent call last):
  File "/workspaces/python/emlearn-micropython/paper/example.py", line 13, in <module>
ValueError: native code in .mpy unsupported
Traceback (most recent call last):
  File "/workspaces/python/emlearn-micropython/paper/create_plot.py", line 119, in <module>
    main()
    ~~~~^^
  File "/workspaces/python/emlearn-micropython/paper/create_plot.py", line 106, in main
    combined, smooth, estimator = train_and_run(sub, class_columns)
                                  ~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^
  File "/workspaces/python/emlearn-micropython/paper/create_plot.py", line 49, in train_and_run
    features = process_data(sensor_data)
  File "/workspaces/python/emlearn-micropython/paper/processing.py", line 80, in process_data
    subprocess.check_output(args)
    ~~~~~~~~~~~~~~~~~~~~~~~^^^^^^
  File "/usr/local/lib/python3.13/subprocess.py", line 472, in check_output
    return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
           ~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
               **kwargs).stdout
               ^^^^^^^^^
  File "/usr/local/lib/python3.13/subprocess.py", line 577, in run
    raise CalledProcessError(retcode, process.args,
                             output=stdout, stderr=stderr)
subprocess.CalledProcessError: Command '['micropython', '/workspaces/python/emlearn-micropython/paper/example.py', '/tmp/tmppsdbb0u8/input.npy', '/tmp/tmppsdbb0u8/output.npy']' returned non-zero exit status 1.
```
