### 2025-12-06 b06a6a2
paper: Document how to use custom executable instead of .mpy modules
References #60
