### 2025-12-06 a9ee34e
make: Add webassembly target
