### 2025-12-06 fabfb54
CI: Add webassembly build
