+++ .github/workflows/build.yaml
@@ -223,6 +223,41 @@ jobs:
         name: emlearn-micropython-build-macos
         path: dist
 
+  build-webassembly:
+    runs-on: ubuntu-24.04
+    env:
+      MPY_DIR: ./micropython
+      MICROPYTHON_BIN: ./micropython/ports/unix/build-nomodules/micropython
+    steps:
+    - uses: actions/checkout@v4
+      with:
+        submodules: true
+    - name: Install OS dependencies
+      run: sudo add-apt-repository universe
+    - uses: actions/setup-python@v5
+      with:
+        python-version: '3.10'
+    - uses: actions/checkout@v4
+      with:
+        repository: jonnor/micropython
+        path: micropython
+        ref: v1.26.1-emlearn
+    - name: Install Python dependencies
+      run: pip install -r requirements.txt
+    - name: Setup MicroPython
+      working-directory: micropython
+      run: |
+        source tools/ci.sh && ci_webassembly_setup
+    - name: Build Webassembly
+      run: |
+        source ${MPY_DIR}/emsdk/emsdk_env.sh
+        make webassembly V=1
+    - name: Archive dist artifacts
+      uses: actions/upload-artifact@v4
+      with:
+        name: emlearn-micropython-build-webassembly
+        path: dist
+
   deploy-pages:
     permissions:
       contents: write
