### 2025-12-06 57d1be1
CI: Try ubuntu-latest for webassembly
