+++ .github/workflows/build.yaml
@@ -224,7 +224,7 @@ jobs:
         path: dist
 
   build-webassembly:
-    runs-on: ubuntu-24.04
+    runs-on: ubuntu-latest
     env:
       MPY_DIR: ./micropython
       MICROPYTHON_BIN: ./micropython/ports/unix/build-nomodules/micropython
