### 2025-12-06 1655d98
CI: Try MicroPython master for webassembly
