### 2025-12-06 a3d19d6
CI: Setup submodules for webassembly
