### 2025-12-06 47f671f
make: Log emcc version and specify variant for webassembly
