### 2025-12-06 ab86293
CI: Try use same emsdk as locally
