### 2025-12-06 5e8b987
CI: Build mpy-cross without emsdk, for webassembly
