### 2025-12-06 85f87ba
CI: Use MicroPython branch with CFLAGS_EXTRA for Webassembly
Otherwise fails on warnings inside emlearn modules
