### 2025-12-06 903d025
CI: Make pages depend on webassembly
