+++ .github/workflows/build.yaml
@@ -265,7 +265,7 @@ jobs:
   deploy-pages:
     permissions:
       contents: write
-    needs: [build-linux, build-arm, build-esp32, build-macos, build-riscv]
+    needs: [build-linux, build-arm, build-esp32, build-macos, build-riscv, build-webassembly]
     runs-on: ubuntu-latest
     
     steps:
