### 2025-12-09 emlearn/emlearn-micropython#60 @swouf
Hi @jonnor, thanks for adding these instructions. Unfortunately, this doesn't work for me still. I have downloaded the micropython binary with the embedded emlearn modules according to your instructions. I also pulled the latest branch as I saw you updated the Makefile. But now, I run into this issue when `python create_plot.py`:

```
missing values 0.0033332962967078145 %
window duration 256 2.56 s
run-cmd micropython /Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/example.py /var/folders/7y/0tz8nnv17hqbfrpz_s5xn7rw0000gn/T/tmpku6v7lxx/input.npy /var/folders/7y/0tz8nnv17hqbfrpz_s5xn7rw0000gn/T/tmpku6v7lxx/output.npy
Traceback (most recent call last):
  File "/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/example.py", line 132, in <module>
  File "/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/example.py", line 91, in process_file
AttributeError: module 'npyfile' has no attribute 'Reader'
Traceback (most recent call last):
  File "/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/create_plot.py", line 119, in <module>
    main()
  File "/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/create_plot.py", line 106, in main
    combined, smooth, estimator = train_and_run(sub, class_columns)
                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/create_plot.py", line 49, in train_and_run
    features = process_data(sensor_data)
               ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/processing.py", line 80, in process_data
    subprocess.check_output(args)
  File "/opt/homebrew/anaconda3/lib/python3.11/subprocess.py", line 466, in check_output
    return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/homebrew/anaconda3/lib/python3.11/subprocess.py", line 571, in run
    raise CalledProcessError(retcode, process.args,
subprocess.CalledProcessError: Command '['micropython', '/Users/jlascala/EPFL/PhD/3️⃣ - 🗺️ Areas/Reviewing/emlearn-micropython/emlearn-micropython/paper/example.py', '/var/folders/7y/0tz8nnv17hqbfrpz_s5xn7rw0000gn/T/tmpku6v7lxx/input.npy', '/var/folders/7y/0tz8nnv17hqbfrpz_s5xn7rw0000gn/T/tmpku6v7lxx/output.npy']' returned non-zero exit status 1.
(base)
```

Apparently, this there is something with the `npyfile` module:
```
AttributeError: module 'npyfile' has no attribute 'Reader'
```

Is it possible that something is missing in micropython binary?
