### 2025-12-09 0851872
paper: Fix curl command for downloading npyfile
curl does not follow redirects by default it seems,
and instead just wrote an empty file
