### 2025-12-10 emlearn/emlearn-micropython#60 @jonnor
That is great to hear that it worked now @swouf  - thank you for the quick iterations on testing and reviewing.

Good point about plotly, that might be missing. Will check and fix that.

Regarding uv - that would not handle the MicroPython part, right? Which is a key dependency here. In the future I want to see MicroPython packaged as a PyPi package - but that work should happen upstream. And that we would have support for .mpy modules on Mac OS, that way not needing a custom `micropython` build.

I see the benefit with uv and I will consider it in the future. I would need to get a bit familiar with the tool myself first - as anything the documentation recommends, I (as a maintainer) need at the very least to be able to debug basics of.
