### 2025-12-11 93c05b5
CI: Deploy only tags - but also to a "latest"
Note: this means that the "master" files are deprecated
