+++ .github/workflows/build.yaml
@@ -279,14 +279,23 @@ jobs:
         mkdir -p dist/
         cp -r artifacts/*/* dist/
 
-    - name: Deploy to Github Pages
-      if: "github.ref_name == 'master' || github.ref_type == 'tag'"
+    - name: Deploy to Github Pages (tag/version)
+      if: "github.ref_type == 'tag'"
       uses: peaceiris/actions-gh-pages@v3
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
         publish_dir: ./dist
         keep_files: true
         destination_dir: builds/${{ github.ref_name }}
+
+    - name: Deploy to Github Pages (latest)
+      if: "github.ref_type == 'tag'"
+      uses: peaceiris/actions-gh-pages@v3
+      with:
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        publish_dir: ./dist
+        keep_files: true
+        destination_dir: builds/latest
 
   docs:
 
