### 2025-12-11 99ed925
CI: Switch to MicroPython 1.27.0
