+++ .github/workflows/build.yaml
@@ -21,9 +21,9 @@ jobs:
         python-version: '3.10'
     - uses: actions/checkout@v4
       with:
-        repository: jonnor/micropython
+        repository: micropython/micropython
         path: micropython
-        ref: v1.26.1-emlearn
+        ref: v1.27.0
     - name: Install Python dependencies
       run: pip install -r requirements.txt
     - name: Setup MicroPython X86
@@ -60,9 +60,9 @@ jobs:
         python-version: '3.10'
     - uses: actions/checkout@v4
       with:
-        repository: jonnor/micropython
+        repository: micropython/micropython
         path: micropython
-        ref: v1.26.1-emlearn
+        ref: v1.27.0
     - name: Install Python dependencies
       run: pip install -r requirements.txt
     - name: Setup MicroPython ARM
@@ -105,9 +105,9 @@ jobs:
         python-version: '3.10'
     - uses: actions/checkout@v4
       with:
-        repository: jonnor/micropython
+        repository: micropython/micropython
         path: micropython
-        ref: v1.26.1-emlearn
+        ref: v1.27.0
     - name: Install Python dependencies
       run: pip install -r requirements.txt
 
@@ -149,9 +149,9 @@ jobs:
         python-version: '3.10'
     - uses: actions/checkout@v4
       with:
-        repository: jonnor/micropython
+        repository: micropython/micropython
         path: micropython
-        ref: v1.26.1-emlearn
+        ref: v1.27.0
     - name: Install Python dependencies
       run: pip install -r requirements.txt
 
@@ -197,9 +197,9 @@ jobs:
         python-version: '3.10'
     - uses: actions/checkout@v4
       with:
-        repository: jonnor/micropython
+        repository: micropython/micropython
         path: micropython
-        ref: v1.26.1-emlearn
+        ref: v1.27.0
     - name: Install Python dependencies
       run: pip install -r requirements.txt
     - name: Setup MicroPython X86
