### 2025-12-11 0a1e741
docs: Bump to MicroPython 1.27
