### 2025-12-11 4ee3e2d
Use latest instead of master for all build URLs
