### 2025-12-11 27a1cd2
cap fixes in paper and bib
