+++ paper/paper.bib
@@ -1,7 +1,7 @@
 
 @misc{micropython,
   author = {George, Damien P and contributors},
-  title = {MicroPython - Python for Microcontrollers},
+  title = {MicroPython - {P}ython for Microcontrollers},
   year = {2014},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -45,8 +45,8 @@
 }
 
 @misc{tflite_micro,
-  author = {Google LLC and contributors},
-  title = {tflite-micro - Tensorflow Lite for Microcontrollers},
+  author = {{Google LLC and contributors}},
+  title = {{t}flite-micro - {T}ensor{F}low {L}ite for {M}icrocontrollers},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -79,8 +79,8 @@
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+  title   = {{SciPy} 1.0: fundamental algorithms for scientific
+            computing in {P}ython},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -160,7 +160,7 @@ author={
   author       = {Ibrahim Abdelkader and
                   Yasser El{-}Sonbaty and
                   Mohamed El{-}Habrouk},
-  title        = {Openmv: {A} Python powered, extensible machine vision camera},
+  title        = {Openmv: {A} {P}ython powered, extensible machine vision camera},
   journal      = {CoRR},
   volume       = {abs/1711.10464},
   year         = {2017},
@@ -173,7 +173,7 @@ author={
 }
 
 @article{karavaev2024tinydecisiontreeclassifier,
-  title={TinyDecisionTreeClassifier: Embedded C++ library for training and applying decision trees on the edge},
+  title={TinyDecisionTreeClassifier: Embedded {C}++ library for training and applying decision trees on the edge},
   author={Karavaev, Aleksei and Hejda, Jan and Kutilek, Patrik and Volf, Petr and Sokol, Marek and Leova, Lydie},
   journal={SoftwareX},
   volume={27},
@@ -302,7 +302,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 }
 
 @article{plauska_performance_2023,
-	title = {Performance Evaluation of C/C++, {MicroPython}, Rust and {TinyGo} Programming Languages on {ESP}32 Microcontroller},
+	title = {Performance {E}valuation of {C}/{C}++, {MicroPython}, {R}ust and {TinyGo} {P}rogramming {L}anguages on {ESP}32 {M}icrocontroller},
 	volume = {12},
 	rights = {http://creativecommons.org/licenses/by/3.0/},
 	issn = {2079-9292},
@@ -323,7 +323,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 }
 
 @inproceedings{ionescu_investigating_2020,
-	title = {Investigating the performance of {MicroPython} and C on {ESP}32 and {STM}32 microcontrollers},
+	title = {Investigating the performance of {MicroPython} and {C} on {ESP}32 and {STM}32 microcontrollers},
 	url = {https://ieeexplore.ieee.org/abstract/document/9292199},
 	doi = {10.1109/SIITME50350.2020.9292199},
[…]
