### 2025-12-13 bb5372a
README: Native modules are supported on Cortex M0
