### 2025-12-13 3234e68
docs: Some notes on developing C modules
