### 2026-01-11 f21d1ad
multiregression: Start of example
