### 2026-01-11 7936c22
multiregression: Initial MicroPython code
