### 2026-01-14 01717c6
trees: Support float leaves for regression
NOTE: requires new emlearn which puts leaf_bits (lf) in .csv export
